+++ packages/stg-definitions-v2/src/constant.ts
@@ -24,6 +24,7 @@ export const DVNS = {
         [EndpointId.AVALANCHE_V2_MAINNET]: '0xa59BA433ac34D2927232918Ef5B2eaAfcF130BA5',
         [EndpointId.BASE_V2_MAINNET]: '0xcd37CA043f8479064e10635020c65FfC005d36f6',
         [EndpointId.BSC_V2_MAINNET]: '0x31F748a368a893Bdb5aBB67ec95F232507601A73',
+        [EndpointId.COREDAO_V2_MAINNET]: '0x7fe673201724925b5c477d4e1a4bd3e954688cf5',
         [EndpointId.EBI_V2_MAINNET]: '0x261150ab73528dbd51573a52917eab243be9729a',
         [EndpointId.ETHEREUM_V2_MAINNET]: '0xa59BA433ac34D2927232918Ef5B2eaAfcF130BA5',
         [EndpointId.FLARE_V2_MAINNET]: '0x9bcd17a654bffaa6f8fea38d19661a7210e22196',
@@ -68,6 +69,7 @@ export const DVNS = {
         [EndpointId.AVALANCHE_V2_MAINNET]: '0x252b234545e154543ad2784c7111eb90406be836',
         [EndpointId.BASE_V2_MAINNET]: '0xcdf31d62140204c08853b547e64707110fbc6680',
         [EndpointId.BSC_V2_MAINNET]: '0xac8de74ce0a44a5e73bbc709fe800406f58431e0',
+        [EndpointId.COREDAO_V2_MAINNET]: '0xe6cd8c2e46ef396df88048449e5b1c75172b40c3',
         [EndpointId.EBI_V2_MAINNET]: '0x97841d4ab18e9a923322a002d5b8eb42b31ccdb5',
         [EndpointId.ETHEREUM_V2_MAINNET]: '0x8fafae7dd957044088b3d0f67359c327c6200d18',
         [EndpointId.FLARE_V2_MAINNET]: '0x8d77d35604a9f37f488e41d1d916b2a0088f82dd',
@@ -99,6 +101,7 @@ export const EXECUTORS = {
         [EndpointId.AVALANCHE_V2_MAINNET]: '0x90E595783E43eb89fF07f63d27B8430e6B44bD9c',
         [EndpointId.BASE_V2_MAINNET]: '0x2CCA08ae69E0C44b18a57Ab2A87644234dAebaE4',
         [EndpointId.BSC_V2_MAINNET]: '0x3ebD570ed38B1b3b4BC886999fcF507e9D584859',
+        [EndpointId.COREDAO_V2_MAINNET]: '0x1785c94d31E3E3Ab1079e7ca8a9fbDf33EEf9dd5',
         [EndpointId.EBI_V2_MAINNET]: '0xc097ab8CD7b053326DFe9fB3E3a31a0CCe3B526f',
         [EndpointId.ETHEREUM_V2_MAINNET]: '0x173272739Bd7Aa6e4e214714048a9fE699453059',
         [EndpointId.FLARE_V2_MAINNET]: '0xcCE466a522984415bC91338c232d98869193D46e',
@@ -255,6 +258,10 @@ export const ASSETS: Record<TokenName, AssetConfig> = {
                 type: StargateType.Pool,
                 address: '0x55d398326f99059fF775485246999027B3197955',
             },
+            [EndpointId.COREDAO_V2_MAINNET]: {
+                type: StargateType.Pool,
+                address: '0x900101d06a7426441ae63e9ab3b9b0f63be145f1',
+            },
             [EndpointId.EBI_V2_MAINNET]: {
                 type: StargateType.Oft,
             },
@@ -367,6 +374,10 @@ export const ASSETS: Record<TokenName, AssetConfig> = {
             },
             [EndpointId.BSC_V2_MAINNET]: {
                 address: '0x8ac76a51cc950d9822d68b83fe1ad97b32cd580d',
+                type: StargateType.Pool,
+            },
+            [EndpointId.COREDAO_V2_MAINNET]: {
+                address: '0xa4151b2b3e269645181dccf2d426ce75fcbdeca9',
                 type: StargateType.Pool,
             },
             [EndpointId.ETHEREUM_V2_MAINNET]: {
@@ -519,6 +530,7 @@ export const OFT_WRAPPER: OftWrapperConfig = {
         [EndpointId.BASE_V2_MAINNET]: {},
         [EndpointId.BLAST_V2_MAINNET]: {},
         [EndpointId.BSC_V2_MAINNET]: {},
+        [EndpointId.COREDAO_V2_MAINNET]: {},
         [EndpointId.EBI_V2_MAINNET]: {},
         [EndpointId.ETHEREUM_V2_MAINNET]: {},
         [EndpointId.ETHERLINK_V2_MAINNET]: {},
@@ -687,6 +699,16 @@ export const REWARDS: RewardsConfig = {
             },
         },
     },
+    [RewardTokenName.CORE]: {
+        name: 'CORE',
+        networks: {
+            //
+            // Mainnet
+            [EndpointId.COREDAO_V2_MAINNET]: {
+                address: '0x0000000000000000000000000000000000000000',
+            },
+        },
+    },
 }
 
 export const NETWORKS: NetworksConfig = {
@@ -714,7 +736,7 @@ export const NETWORKS: NetworksConfig = {
         safeConfig: {
             safeAddress: '0x5d3917b47e963ec703ed66da6637c701365ff500',
             safeUrl:
-                'https://gnosis.lzdevnet.org/transactions/queue?safe=astar:0x5d3917b47e963ec703ed66da6637c701365ff500',
+                'https://astar-tx.lzdevnet.org/',
         },
     },
     [EndpointId.AURORA_V2_MAINNET]: {
@@ -773,7 +795,7 @@ export const NETWORKS: NetworksConfig = {
     [EndpointId.BLAST_V2_MAINNET]: {
         safeConfig: {
             safeAddress: '0xc53329FD24f3a446b7c3a804Ebc53515c0244012',
-            safeUrl: 'https://gn.lzdevnet.org/app/blast:0xc53329FD24f3a446b7c3a804Ebc53515c0244012/transactions/queue',
+            safeUrl: 'https://blast-tx.lzdevnet.org/',
         },
     },
     [EndpointId.BSC_V2_MAINNET]: {
@@ -793,17 +815,36 @@ export const NETWORKS: NetworksConfig = {
             safeUrl: 'https://safe-transaction-bsc.safe.global/',
         },
     },
+    [EndpointId.COREDAO_V2_MAINNET]: {
+        creditMessaging: {
+            ...DEFAULT_CREDIT_MESSAGING_NETWORK_CONFIG,
+            requiredDVNs: [DVNS.NETHERMIND[EndpointId.COREDAO_V2_MAINNET], DVNS.STG[EndpointId.COREDAO_V2_MAINNET]],
+            executor: EXECUTORS.LZ_LABS[EndpointId.COREDAO_V2_MAINNET],
+        },
+        tokenMessaging: {
+            ...DEFAULT_TOKEN_MESSAGING_NETWORK_CONFIG,
+            requiredDVNs: [DVNS.NETHERMIND[EndpointId.COREDAO_V2_MAINNET], DVNS.STG[EndpointId.COREDAO_V2_MAINNET]],
+            executor: EXECUTORS.LZ_LABS[EndpointId.COREDAO_V2_MAINNET],
+            nativeDropAmount: parseEther('0.1').toBigInt(),
+        },
+        safeConfig: {
+            safeAddress: '0x43303706f215A53220291F0B8a896BCDA5EB709E',
+            safeUrl: 'https://core-tx.lzdevnet.org/',
+        },
+    },
     [EndpointId.EBI_V2_MAINNET]: {
         creditMessaging: {
             ...DEFAULT_CREDIT_MESSAGING_NETWORK_CONFIG,
             requiredDVNs: [DVNS.NETHERMIND[EndpointId.EBI_V2_MAINNET], DVNS.STG[EndpointId.EBI_V2_MAINNET]],
             executor: EXECUTORS.LZ_LABS[EndpointId.EBI_V2_MAINNET],
+            confirmations: BigInt(50),
         },
         tokenMessaging: {
             ...DEFAULT_TOKEN_MESSAGING_NETWORK_CONFIG,
             requiredDVNs: [DVNS.NETHERMIND[EndpointId.EBI_V2_MAINNET], DVNS.STG[EndpointId.EBI_V2_MAINNET]],
             executor: EXECUTORS.LZ_LABS[EndpointId.EBI_V2_MAINNET],
             nativeDropAmount: parseEther('0.00003').toBigInt(),
+            confirmations: BigInt(50),
         },
         safeConfig: {
             safeAddress: '0xb93Aa694A3De8662E1ca9aD0C811440E48cDFe5E',
@@ -843,13 +884,13 @@ export const NETWORKS: NetworksConfig = {
         safeConfig: {
             safeAddress: '0x757A404a44C9fC75136e8901E561ac2bcc9FCE8D',
             safeUrl:
-                'https://gnosis.lzdevnet.org/transactions/queue?safe=etlk:0x757A404a44C9fC75136e8901E561ac2bcc9FCE8D',
+                'https://etherlink-tx.lzdevnet.org/',
         },
     },
     [EndpointId.FANTOM_V2_MAINNET]: {
         safeConfig: {
             safeAddress: '0x2351BBCb7cF7Ee9D18AF2Be0d106BFc5D47A9E85',
-            safeUrl: 'https://gn.lzdevnet.org/app/ftm:0x2351BBCb7cF7Ee9D18AF2Be0d106BFc5D47A9E85/transactions/queue',
+            safeUrl: 'https://fantom-tx.lzdevnet.org/',
         },
     },
     [EndpointId.FLARE_V2_MAINNET]: {
@@ -885,7 +926,7 @@ export const NETWORKS: NetworksConfig = {
         safeConfig: {
             safeAddress: '0x62B5F0B624301A1F5C0DD998A40Ea7297B26FB90',
             safeUrl:
-                'https://gn.lzdevnet.org/app/fraxtal:0x62B5F0B624301A1F5C0DD998A40Ea7297B26FB90/transactions/queue',
+                'https://fraxtal-tx.lzdevnet.org/',
         },
     },
     [EndpointId.GRAVITY_V2_MAINNET]: {
@@ -983,7 +1024,7 @@ export const NETWORKS: NetworksConfig = {
     [EndpointId.MANTA_V2_MAINNET]: {
         safeConfig: {
             safeAddress: '0x3f0DFccF4f7BBc0ed52A212e4d981435a7f27Cc6',
-            safeUrl: 'https://gn.lzdevnet.org/app/manta:0x3f0DFccF4f7BBc0ed52A212e4d981435a7f27Cc6/transactions/queue',
+            safeUrl: 'https://manta-tx.lzdevnet.org/',
         },
     },
     [EndpointId.MANTLE_V2_MAINNET]: {
@@ -1024,28 +1065,28 @@ export const NETWORKS: NetworksConfig = {
     [EndpointId.MODE_V2_MAINNET]: {
         safeConfig: {
             safeAddress: '0x9BD292C0f5D95686481A8af9b8694654B25aE1AC',
-            safeUrl: 'https://gn.lzdevnet.org/app/mode:0x9BD292C0f5D95686481A8af9b8694654B25aE1AC/transactions/queue',
+            safeUrl: 'https://mode-tx.lzdevnet.org/',
         },
     },
     [EndpointId.MOONBEAM_V2_MAINNET]: {
         safeConfig: {
             safeAddress: '0x40533743FC0F3cCb01ca2196d45dd7958dc89f89',
             safeUrl:
-                'https://gnosis.lzdevnet.org/transactions/queue?safe=mbeam:0x40533743FC0F3cCb01ca2196d45dd7958dc89f89',
+                'https://moonbeam-tx.lzdevnet.org/',
         },
     },
     [EndpointId.MOONRIVER_V2_MAINNET]: {
         safeConfig: {
             safeAddress: '0xBAc08c612a791033BC20D991FB9b1892Cb49A39f',
             safeUrl:
-                'https://gnosis.lzdevnet.org/transactions/queue?safe=mriver:0xBAc08c612a791033BC20D991FB9b1892Cb49A39f',
+                'https://moonriver-tx.lzdevnet.org/',
         },
     },
     [EndpointId.OPBNB_V2_MAINNET]: {
         safeConfig: {
             safeAddress: '0xD6578c1C35ee901d01D99e17593E25B13994090b',
             safeUrl:
-                'https://gnosis.lzdevnet.org/transactions/history?safe=obnb:0xD6578c1C35ee901d01D99e17593E25B13994090b',
+                'https://opbnb-tx.lzdevnet.org/',
         },
     },
     [EndpointId.OPTIMISM_V2_MAINNET]: {
@@ -1134,7 +1175,7 @@ export const NETWORKS: NetworksConfig = {
         safeConfig: {
             safeAddress: '0x3ae59e4cffaad28e6588a269e2142e4a434d5a94',
             safeUrl:
-                'https://gnosis.lzdevnet.org/transactions/queue?safe=shimmer:0x3ae59e4cffaad28e6588a269e2142e4a434d5a94',
+                'https://shimmer-tx.lzdevnet.org/',
         },
     },
     [EndpointId.SEI_V2_MAINNET]: {
@@ -1228,7 +1269,7 @@ export const NETWORKS: NetworksConfig = {
         safeConfig: {
             safeAddress: '0x64a77dD82517cC8023a52D27f4c167439bDeF5B9',
             safeUrl:
-                'https://gnosis.lzdevnet.org/transactions/queue?safe=astrzk:0x64a77dD82517cC8023a52D27f4c167439bDeF5B9',
+                'https://zkatana-tx.lzdevnet.org/',
         },
     },
     [EndpointId.ZKCONSENSYS_V2_MAINNET]: {
@@ -1257,14 +1298,14 @@ export const NETWORKS: NetworksConfig = {
     [EndpointId.ZKPOLYGON_V2_MAINNET]: {
         safeConfig: {
             safeAddress: '0x5D3917b47E963eC703eD66Da6637C701365fF500',
-            safeUrl: 'https://app.safe.global/transactions/queue?safe=zkevm:0x5D3917b47E963eC703eD66Da6637C701365fF500',
+            safeUrl: 'https://safe-transaction-zkevm.safe.global/',
         },
     },
     [EndpointId.ZKSYNC_V2_MAINNET]: {
         safeConfig: {
             safeAddress: '0x026756AB43866eCd92289663E91CCa8afb20414B',
             safeUrl:
-                'https://app.safe.global/transactions/queue?safe=zksync:0x026756AB43866eCd92289663E91CCa8afb20414B',
+                'https://safe-transaction-zksync.safe.global/',
         },
     },