+++ packages/stg-definitions-v2/src/types.ts
@@ -56,6 +56,7 @@ export enum RewardTokenName {
     wKAVA = 'wKAVA',
     AURORA = 'AURORA',
     SEI = 'SEI',
+    CORE = 'CORE',
 }
 
 export type RewardsConfig = Record<RewardTokenName, RewardsTokenConfig>
@@ -91,6 +92,7 @@ export interface CreditMessagingNetworkConfig {
     sendCreditGasLimit: bigint
     requiredDVNs?: string[]
     executor?: string
+    confirmations?: bigint
 }
 
 export interface TokenMessagingNetworkConfig {
@@ -103,6 +105,7 @@ export interface TokenMessagingNetworkConfig {
     requiredDVNs?: string[]
     executor?: string
     queueCapacity: number
+    confirmations?: bigint
 }
 
 export interface OftWrapperConfig {